moved app colors from 'theme' to 'common';
