+++ lib/presentation/pages/comments_page/widgets/comment_field.dart
@@ -1,5 +1,5 @@
+import 'package:blogify_flutter_main/common/app_colors.dart';
 import 'package:blogify_flutter_main/common/app_dimensions.dart';
-import 'package:blogify_flutter_main/presentation/theme/app_colors.dart';
 import 'package:flutter/material.dart';
 
 class CommentField extends StatelessWidget {
