+++ lib/presentation/pages/home_page/widgets/home_search_bar.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
 
+import '../../../../common/app_colors.dart';
 import '../../../../data/mock_storage/global_mock_storage_provider.dart';
-import '../../../theme/app_colors.dart';
 import '../notifiers/search_column_notifier.dart';
 import 'menu_item.dart';
