+++ lib/presentation/pages/home_page/widgets/post_card.dart
@@ -1,6 +1,6 @@
+import 'package:blogify_flutter_main/common/app_colors.dart';
 import 'package:blogify_flutter_main/common/app_dimensions.dart';
 import 'package:blogify_flutter_main/domain/entities/post_entity.dart';
-import 'package:blogify_flutter_main/presentation/theme/app_colors.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../utils/intl_formatter.dart';
